requirements.txt
extdirect.django